pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "edulynx"
version = "0.1.0"
description = "A python FastApi Application to manage courses."
readme = "README.md"
authors = ["Vaibhav Vikas <vbhvvikas@gmail.com>"]
license = "GNU GPL v3.0"
repository = "https://github.com/vaibhavvikas/edulynx"
homepage = "https://github.com/vaibhavvikas/edulynx"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"edulynx" = "edulynx.__main__:start"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.7.0"}
rich = "^12.0.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
pymongo = "^4.3.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
darglint = "^1.8.1"
isort = {extras = ["colors"], version = "^5.12.0"}
mypy = "^1.2.0"
pre-commit = "^3.3.1"
pydocstyle = "^6.3.0"
pylint = "^2.17.4"
pytest = "^7.3.1"
pyupgrade = "^3.4.0"
safety = {version = "^2.4.0b1", allow-prereleases = true}
coverage = "^7.2.5"
coverage-badge = "^1.1.0"
pytest-html = "^3.2.0"
pytest-cov = "^4.0.0"
black = {version = "^23.3.0", allow-prereleases = true}
httpx = "^0.24.0"

[tool.black]
# https://github.com/psf/black
target-version = ["py311"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 311
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.11
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]

[coverage.paths]
source = "edulynx"

[coverage.run]
branch = true

[coverage.report]
fail_under = 50
show_missing = true
